feat: huge performance refactor
new validator operates on new param class per generator/strategy. this allows for validation of params set either in template or dynamically in code

all generators and strategies are broken out into seperate class files. it was getting too big to manage, and if in the future we do a plugin system, we will have easy entry points to do so.